global variables definition moved from actors to const.cpp
