+++ core/vm/actor/builtin/types/CMakeLists.txt
@@ -3,10 +3,7 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
-add_subdirectory(market)
 add_subdirectory(miner)
 add_subdirectory(multisig)
-add_subdirectory(payment_channel)
 add_subdirectory(reward)
 add_subdirectory(storage_power)
-add_subdirectory(verified_registry)
