--- core/vm/actor/builtin/types/market/CMakeLists.txt
@@ -1,11 +1,0 @@
-#
-# Copyright Soramitsu Co., Ltd. All Rights Reserved.
-# SPDX-License-Identifier: Apache-2.0
-#
-
-add_library(market_types
-    policy.cpp
-    )
-target_link_libraries(market_types
-    const
-    )
