+++ core/vm/actor/builtin/types/miner/CMakeLists.txt
@@ -5,7 +5,6 @@
 
 add_library(miner_types
     deadline_info.cpp
-    policy.cpp
     )
 target_link_libraries(miner_types
     actor
