--- core/vm/actor/builtin/types/payment_channel/CMakeLists.txt
@@ -1,11 +1,0 @@
-#
-# Copyright Soramitsu Co., Ltd. All Rights Reserved.
-# SPDX-License-Identifier: Apache-2.0
-#
-
-add_library(payment_channel_types
-    policy.cpp
-    )
-target_link_libraries(payment_channel_types
-    const
-    )
