--- core/vm/actor/builtin/types/verified_registry/CMakeLists.txt
@@ -1,11 +1,0 @@
-#
-# Copyright Soramitsu Co., Ltd. All Rights Reserved.
-# SPDX-License-Identifier: Apache-2.0
-#
-
-add_library(verified_registry_types
-    policy.cpp
-    )
-target_link_libraries(verified_registry_types
-    Boost::boost
-    )
