+++ core/vm/actor/builtin/types/verified_registry/policy.hpp
@@ -5,14 +5,10 @@
 
 #pragma once
 
-#include "const.hpp"
 #include "primitives/types.hpp"
 
 namespace fc::vm::actor::builtin::types::verified_registry {
   using primitives::StoragePower;
 
   extern StoragePower kMinVerifiedDealSize;
-
-  void setPolicy(const StoragePower &minVerifiedDealSize);
-
 }  // namespace fc::vm::actor::builtin::types::verified_registry
