+++ core/vm/actor/builtin/v0/market/CMakeLists.txt
@@ -14,5 +14,4 @@ target_link_libraries(market_actor_v0
     balance_table
     cid_key
     piece
-    market_types
     )
