+++ core/vm/actor/builtin/v0/miner/CMakeLists.txt
@@ -17,5 +17,4 @@ target_link_libraries(miner_actor_v0
     rle_bitset
     sector
     uvarint_key
-    miner_types
     )
