+++ core/vm/actor/builtin/v0/payment_channel/CMakeLists.txt
@@ -13,5 +13,4 @@ target_link_libraries(payment_channel_actor_v0
     array
     outcome
     signature
-    payment_channel_types
     )
