+++ core/vm/actor/builtin/v0/verified_registry/CMakeLists.txt
@@ -11,5 +11,4 @@ add_library(verified_registry_actor_v0
 target_link_libraries(verified_registry_actor_v0
     actor
     outcome
-    verified_registry_types
     )
